Structural and other minor changes
-Added modules folder, moved func.js (now util.js) and server list there)
-Changed as many CamelCase variables as I could find into snake_case (snake case for variables, camel case for functions)
-Minor changes to how messages are deleted
-Removed index.js requiring util.js and util.js requiring index.js
